+++ frontend/components/Article/Article.module.css
@@ -4,7 +4,7 @@
   justify-content: center;
   flex-wrap: wrap;
   max-width: 800px;
-  margin-top: 3rem;
+  margin:3rem;
 }
 
 .card {
