+++ frontend/components/Footer/Footer.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Image from 'next/image';
+import {BgWrap} from 'components/Footer/FooterStyle'
+const Footer = () => {
+    return (
+            <BgWrap>
+                <div style={{color:'#fff'}}>어디갔냐</div>
+            </BgWrap>
+    )
+}
+
+export default Footer
