+++ frontend/components/Footer/FooterStyle.ts
@@ -1,0 +1,11 @@
+import { css } from '@emotion/react'
+import styled from '@emotion/styled'
+
+export const BgWrap=styled.div`
+    width:100vw;
+    min-height: 200px;
+    background:url("/static/footer-bg.png");
+    background-position: center;
+    background-size: cover;
+`;
+
