+++ frontend/styles/commonStyles.js
@@ -1,0 +1,9 @@
+import { css } from '@emotion/react'
+import styled from '@emotion/styled'
+
+export const ImgBox=styled.div`
+    width:100%;
+    height:auto;
+    background:pink;
+    position:fixed;
+`;
